+++ testing/behavioural/src/selection/group-data.ts
